refactor: move PasswordEntropyCalculator to services/
Pure utility function belongs in services/ not components/.
Update imports in PasswordStrength.jsx and test file.
